scale_factor bug修复

+++ anomalib/models/patchcore/anomaly_map.py
@@ -37,7 +37,7 @@ class AnomalyMapGenerator(nn.Module):
             torch.Tensor: Map of the pixel-level anomaly scores
         """
         # scale_factor代替size
-        scale_factor = [self.input_size[0] / patch_scores.shape[-2], self.input_size[1] / patch_scores.shape[-1]]
+        scale_factor = [int(self.input_size[0] / patch_scores.shape[-2]), int(self.input_size[1] / patch_scores.shape[-1])]
         anomaly_map = F.interpolate(patch_scores, scale_factor=scale_factor)
         anomaly_map = self.blur(anomaly_map)
